+++ src/nitrokey/trussed/_bootloader/nrf52_upload/dfu/init_packet_pb.py
@@ -35,19 +35,104 @@
 # SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #
 
+from enum import Enum
+from typing import Any, List, Optional
+
 from . import dfu_cc_pb2 as pb
 
 
-class InitPacketPB:
-    def __init__(self, from_bytes: bytes) -> None:
-        # construct from a protobuf string/buffer
-        self.packet = pb.Packet()
-        self.packet.ParseFromString(from_bytes)
+class SigningTypes(Enum):
+    ECDSA_P256_SHA256 = pb.ECDSA_P256_SHA256
+    ED25519 = pb.ED25519
 
-        if self.packet.HasField("signed_command"):
-            self.init_command = self.packet.signed_command.command.init
+
+class CommandTypes(Enum):
+    INIT = pb.INIT
+
+
+class HashTypes(Enum):
+    NONE = pb.NO_HASH
+    CRC = pb.CRC
+    SHA128 = pb.SHA128
+    SHA256 = pb.SHA256
+    SHA512 = pb.SHA512
+
+
+class DFUType(Enum):
+    APPLICATION = pb.APPLICATION
+    SOFTDEVICE = pb.SOFTDEVICE
+    BOOTLOADER = pb.BOOTLOADER
+    SOFTDEVICE_BOOTLOADER = pb.SOFTDEVICE_BOOTLOADER
+    EXTERNAL_APPLICATION = pb.EXTERNAL_APPLICATION
+
+
+class ValidationTypes(Enum):
+    NO_VALIDATION = pb.NO_VALIDATION
+    VALIDATE_GENERATED_CRC = pb.VALIDATE_GENERATED_CRC
+    VALIDATE_GENERATED_SHA256 = pb.VALIDATE_SHA256
+    VALIDATE_ECDSA_P256_SHA256 = pb.VALIDATE_ECDSA_P256_SHA256
+
+
+class InitPacketPB:
+    def __init__(
+        self,
+        from_bytes: Optional[bytes] = None,
+        hash_bytes: Optional[bytes] = None,
+        hash_type: Optional[HashTypes] = None,
+        boot_validation_type: List[Any] = [],  # noqa: B006
+        boot_validation_bytes: List[Any] = [],  # noqa: B006
+        dfu_type: Any = None,
+        is_debug: bool = False,
+        fw_version: int = 0xFFFFFFFF,
+        hw_version: int = 0xFFFFFFFF,
+        sd_size: int = 0,
+        app_size: int = 0,
+        bl_size: int = 0,
+        sd_req: Any = None,
+    ) -> None:
+        if from_bytes is not None:
+            # construct from a protobuf string/buffer
+            self.packet = pb.Packet()
+            self.packet.ParseFromString(from_bytes)
+
+            if self.packet.HasField("signed_command"):
+                self.init_command = self.packet.signed_command.command.init
+            else:
+                self.init_command = self.packet.command.init
+
         else:
-            self.init_command = self.packet.command.init
+            # construct from input variables
+            if not sd_req:
+                sd_req = [0xFFFE]  # Set to default value
+            self.packet = pb.Packet()
+
+            boot_validation = []
+            for i, x in enumerate(boot_validation_type):
+                boot_validation.append(
+                    pb.BootValidation(type=x.value, bytes_=boot_validation_bytes[i])
+                )
+
+            # By default, set the packet's command to an unsigned command
+            # If a signature is set (via set_signature), this will get overwritten
+            # with an instance of SignedCommand instead.
+            self.packet.command.op_code = pb.INIT
+
+            self.init_command = pb.InitCommand()
+            self.init_command.hash.hash_type = (
+                hash_type.value if hash_type else HashTypes.NONE.value
+            )
+            self.init_command.type = dfu_type.value
+            self.init_command.hash.hash = hash_bytes if hash_bytes else b""
+            self.init_command.is_debug = is_debug
+            self.init_command.fw_version = fw_version
+            self.init_command.hw_version = hw_version
+            self.init_command.sd_req.extend(list(set(sd_req)))
+            self.init_command.sd_size = sd_size
+            self.init_command.bl_size = bl_size
+            self.init_command.app_size = app_size
+
+            self.init_command.boot_validation.extend(boot_validation)
+            self.packet.command.init.CopyFrom(self.init_command)
 
         self._validate()
 
@@ -81,8 +166,27 @@ class InitPacketPB:
                 "Invalid range of firmware argument. [0 - 0xffffffff] is valid range"
             )
 
+    def _is_valid(self) -> bool:
+        try:
+            self._validate()
+        except RuntimeError:
+            return False
+
+        return self.packet.signed_command.signature is not None
+
+    def get_init_packet_pb_bytes(self) -> bytes:
+        return self.packet.SerializeToString()
+
     def get_init_command_bytes(self) -> bytes:
         return self.init_command.SerializeToString()
+
+    def set_signature(self, signature: bytes, signature_type: SigningTypes) -> None:
+        new_packet = pb.Packet()
+        new_packet.signed_command.signature = signature
+        new_packet.signed_command.signature_type = signature_type.value
+        new_packet.signed_command.command.CopyFrom(self.packet.command)
+
+        self.packet = new_packet
 
     def __str__(self) -> str:
         return str(self.init_command)