+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - `nitrokey.nk3.secrets_app`: Fix discarded authentication when using `SecretsApp` over CCID
 - Bump minimum Python version to 3.11.
 - `nitrokey.trussed`: Add the `inner_checksum` field to `FirmwareMetadata` with the checksum of the unsigned firmware image.
+- `nitrokey.trussed.nrfutils`: Add support for NRF52 package signing and upload
 
 [All Changes](https://github.com/Nitrokey/nitrokey-sdk-py/compare/v0.5.0-rc.2...HEAD)